Added Increase counter functionality
