+++ app/build.gradle
@@ -6,7 +6,7 @@ apply plugin: "androidx.navigation.safeargs.kotlin"
 
 android {
     compileSdkVersion 29
-    buildToolsVersion "29.0.0"
+    buildToolsVersion "29.0.2"
     defaultConfig {
         applicationId "com.example.justatimerandcounter"
         minSdkVersion 19
@@ -40,4 +40,6 @@ dependencies {
     // Navigation dependency
     implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
     implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
+    // ViewModel dependencies
+    implementation "androidx.lifecycle:lifecycle-extensions:$lifecycle_version"
 }
