+++ app/src/main/java/com/example/justatimerandcounter/screens/counter/CounterViewModel.kt
@@ -1,0 +1,21 @@
+package com.example.justatimerandcounter.screens.counter
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class CounterViewModel: ViewModel() {
+
+    // Current counter value
+    private val _score = MutableLiveData<Int>()
+    val score: LiveData<Int>
+        get() = _score
+
+    init {
+        _score.value = 0
+    }
+
+    fun onCount() {
+        _score.value = _score.value?.plus(1)
+    }
+}
