+++ app/src/main/res/layout/fragment_counter.xml
@@ -3,7 +3,14 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.example.justatimerandcounter.CounterFragment">
+    tools:context="com.example.justatimerandcounter.screens.counter.CounterFragment">
+
+    <data>
+        <variable
+            name="counterViewModel"
+            type="com.example.justatimerandcounter.screens.counter.CounterViewModel" />
+    </data>
+
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
@@ -30,6 +37,7 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/counter_number" />
+            app:layout_constraintTop_toBottomOf="@+id/counter_number"
+            android:onClick="@{() -> counterViewModel.onCount()}"/>
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
