+++ build.gradle
@@ -5,7 +5,7 @@ buildscript {
         version_core = "1.0.1"
         kotlin_version = '1.3.50'
         nav_version = "2.1.0"
-        version_navigation = "1.0.0"
+        lifecycle_version = "2.1.0"
     }
     repositories {
         google()
